base/requirements.txt
#Third party package requirements for OpenEats 2


coreapi-cli==1.0.6
Django==1.10
django-imagekit==4.0.1
djangorestframework==3.4.6
djangorestframework-bulk==0.2.1
gunicorn==19.6.0
requests==2.11.1
six==1.10.0
django-filter==0.15.0
django-extensions==1.7.2
Pillow==2.7.0
django-cors-headers==1.3.1
MySQL-python==1.2.5
mock==2.0.0
pytz==2016.10

# graphene-django==1.3
git+https://github.com/graphql-python/graphene-django.git@2929d0866c800ae6efa8d35bb40548940a70c31f
graphene==1.4.1
